refactor(value): split Tag/TaggedValue into a submodule
Final pass on the value.rs decomposition (follow-up to #105, #106):

- value/tag.rs — Tag, TaggedValue, MaybeTag, nobang, check_for_tag, the
  pub(crate) TAGGED_VALUE_FIELD_* constants, and the tag-preserving serde
  access types (599 lines).

Re-exported from value.rs so existing paths resolve unchanged:
- public: `pub use tag::{MaybeTag, Tag, TaggedValue, check_for_tag, nobang}`
- crate-internal: `pub(crate) use tag::{TAGGED_VALUE_FIELD_*,
  TagPreservingMapAccess}` (serde_impl.rs uses the constants via super::,
  de.rs uses crate::value::TagPreservingMapAccess).

The one non-mechanical change: TaggedValue.value is now pub(crate)
(was a same-module private field that Value::untag and friends move out
of). It is not part of the public API, so semver is unaffected.

value.rs drops 2,304 -> 1,717 lines (4,637 -> 1,717 cumulative, -63%).

Verified: build, fmt, clippy -D warnings, and `cargo test --all-features`
(527 lib + all integration suites + doctests) green.



---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co
